@@ -6,8 +6,12 @@ import {
   tick,
   waitForAsync
 } from '@angular/core/testing';
-import { UntypedFormBuilder, UntypedFormControl, ReactiveFormsModule } from '@angular/forms';
-
+import {
+  FormsModule,
+  ReactiveFormsModule,
+  UntypedFormBuilder,
+  UntypedFormControl
+} from '@angular/forms';
 import { toInputSafeName } from '../_helpers';
 import { createMockPipe } from '../_mocked';
 import { DimensionName } from '../_models';
@@ -30,15 +34,13 @@ describe('FilterComponent', () => {
 
   const emptyOptions = { options: [], hasMore: false };
 
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        imports: [ReactiveFormsModule],
-        declarations: [FilterComponent, createMockPipe('renameApiFacet')],
-        schemas: [CUSTOM_ELEMENTS_SCHEMA]
-      }).compileComponents();
-    })
-  );
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [FormsModule, ReactiveFormsModule],
+      declarations: [FilterComponent, createMockPipe('renameApiFacet')],
+      schemas: [CUSTOM_ELEMENTS_SCHEMA]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(FilterComponent);
