@@ -1,9 +1,9 @@
 import { CUSTOM_ELEMENTS_SCHEMA, ElementRef } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import {
+  ReactiveFormsModule,
   UntypedFormBuilder,
   UntypedFormControl,
-  ReactiveFormsModule,
   ValidationErrors
 } from '@angular/forms';
 import { validateDateGeneric } from '../_helpers';
