@@ -10,7 +10,7 @@ const config = {
     ]
   },
   resolve: {
-    extensions: ['.tsx', '.ts', '.js']
+    extensions: ['.tsx', '.ts', '.js'], modules: ['src', 'node_modules']
   }
 };
 
